Move amici.jax to amici.exporters.jax and amici.sim.jax
Related to #3041.
