+++ python/sdist/amici/exporters/sundials/__init__.py
@@ -1,0 +1,1 @@
+"""Code generation for models for simulation with SUNDIALS solvers."""
